fix role permissions in views

+++ resources/views/admin/areas/index.blade.php
@@ -5,12 +5,13 @@
 
   <div class="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center pt-3 pb-2 mb-3 border-bottom">
     <h1 class="h2">INDEX ADMIN AREA</h1>
-    <!--<p id="demo">a</p>-->
-    <div class="btn-toolbar mb-2 mb-md-0">
-      <div class="btn-group mr-2">        
-        <a href="/admin/areas/create" class="btn btn-primary btn-lg active" role="button" aria-pressed="true">Crear Area</a>
-      </div>     
-    </div>
+    @if(Auth::user()->role_id == 1)
+      <div class="btn-toolbar mb-2 mb-md-0">
+        <div class="btn-group mr-2">        
+          <a href="/admin/areas/create" class="btn btn-primary btn-lg active" role="button" aria-pressed="true">Crear Area</a>
+        </div>     
+      </div>
+    @endif
   </div>
  
   @if(count($areas))
@@ -19,8 +20,10 @@
       <thead class="thead-dark">
         <tr>          
           <th scope="col">Id</th>
-          <th scope="col">Nombre</th>          
-          <th scope="col">Opciones</th>          
+          <th scope="col">Nombre</th>       
+          @if(Auth::user()->role_id == 1)   
+            <th scope="col">Opciones</th>          
+          @endif
         </tr>
       </thead>
       <tbody>
@@ -29,12 +32,14 @@
         <tr>          
           <td>{{$area->id}}</td>
           <td>{{$area->name}}</td>         
-          <td>
-            <a href= "{{route('areas.show', $area -> id) }}"> Ver </a> &nbsp;
-            <a href= "{{route('areas.edit', $area -> id) }}"> Editar </a> &nbsp;
-            <meta name="csrf-token" content="{{ csrf_token() }}">
-            <a href="{{ route('areas.destroy', $area->id) }}" data-method="delete" class="jquery-postback">Delete</a>
-          </td>          
+          @if(Auth::user()->role_id == 1)
+            <td>
+              <a href= "{{route('areas.show', $area -> id) }}"> Ver </a> &nbsp;
+              <a href= "{{route('areas.edit', $area -> id) }}"> Editar </a> &nbsp;
+              <meta name="csrf-token" content="{{ csrf_token() }}">
+              <a href="{{ route('areas.destroy', $area->id) }}" data-method="delete" class="jquery-postback">Delete</a>
+            </td>          
+          @endif
         </tr>
       @endforeach
