+++ resources/views/admin/users/index.blade.php
@@ -4,12 +4,14 @@
 @section("contenido") 
 
   <div class="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center pt-3 pb-2 mb-3 border-bottom">
-    <h1 class="h2">INDEX ADMIN USER</h1>    
-    <div class="btn-toolbar mb-2 mb-md-0">
-      <div class="btn-group mr-2">        
-        <a href="/admin/users/create" class="btn btn-primary btn-lg active" role="button" aria-pressed="true">Crear Usuario</a>
-      </div>     
-    </div>
+    <h1 class="h2">INDEX ADMIN USER</h1>
+    @if(Auth::user()->role_id == 1)    
+      <div class="btn-toolbar mb-2 mb-md-0">
+        <div class="btn-group mr-2">        
+          <a href="/admin/users/create" class="btn btn-primary btn-lg active" role="button" aria-pressed="true">Crear Usuario</a>
+        </div>     
+      </div>
+    @endif
   </div>
  
   @if(count($usuarios))
@@ -21,7 +23,9 @@
           <th scope="col">Nombre</th>
           <th scope="col">Email</th>          
           <th scope="col">Rol</th>
-          <th scope="col">Opciones</th>          
+          @if(Auth::user()->role_id == 1)
+            <th scope="col">Opciones</th>          
+          @endif
         </tr>
       </thead>
       <tbody>
@@ -32,12 +36,14 @@
           <td>{{$usuario->name}}</td>
           <td>{{$usuario->email}}</td>          
           <td>{{$usuario->role->name}}</td>
-          <td>
-            <a href= "{{route('users.show', $usuario -> id) }}"> Ver </a> &nbsp;
-            <a href= "{{route('users.edit', $usuario -> id) }}"> Editar </a> &nbsp;
-            <meta name="csrf-token" content="{{ csrf_token() }}">
-            <a href="{{ route('users.destroy', $usuario->id) }}" data-method="delete" class="jquery-postback">Delete</a>
-          </td>          
+          @if(Auth::user()->role_id == 1)
+            <td>
+              <a href= "{{route('users.show', $usuario -> id) }}"> Ver </a> &nbsp;
+              <a href= "{{route('users.edit', $usuario -> id) }}"> Editar </a> &nbsp;
+              <meta name="csrf-token" content="{{ csrf_token() }}">
+              <a href="{{ route('users.destroy', $usuario->id) }}" data-method="delete" class="jquery-postback">Delete</a>
+            </td>          
+          @endif
         </tr>
       @endforeach
